Rewrite INGAMEPORT to use the Multimedia API provided on all Win32 systems.
This makes it no longer crash on NT-based systems.
It also makes it faster on Win95-based systems.

It also ignores the "portid" parameter, which I think is an acceptable regressions.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@64 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ trunk/logo/src/func.h
@@ -10,7 +10,7 @@ void Myoutportb(short portid, unsigned char value);
 unsigned char Myinportb(short portid);
 void Myoutport(short portid, short value);
 short Myinport(short portid);
-short Myingameport (short portid, short mask);
+int Myingameport (short portid, short mask);
 
 extern void gr_mode(void);             /*routine                              */
 extern void ibm_prepare_to_draw_turtle(void);
